+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The following is a history of the changes made to this project.
 * Modified the airplanes.live client installation script so it utilizes the new logging functions.
 * Modified the FlightRadar24 client installation script so it utilizes the new logging functions.
 * Modified the Fly Italy ADS-B client installation script so it utilizes the new logging functions.
+* Modified the OpenSky Network client installation script so it utilizes the new logging functions.
+* Modified the PlaneFinder client installation script so it utilizes the new logging functions.
 * Added the --version and -v arguments in order to display the current project version.
 
 ## v2.8.4 *(July 13th, 2024)* :rooster: